create useWorkflowCommandMenu hook
